refactor(io): move stores into files/ + remove re-export shims (#366 Phase 5)
Final phase of #366. All workspace file I/O now lives under a single
directory — server/utils/files/.

### Store helpers moved (git mv, history preserved)

- image-store.ts → files/image-store.ts
- markdown-store.ts → files/markdown-store.ts
- spreadsheet-store.ts → files/spreadsheet-store.ts

Internal imports updated (../workspace/paths → ../../workspace/paths,
./fs → ./safe).

### Re-export shims deleted

- server/utils/file.ts (9 consumers migrated to files/index.js)
- server/utils/fs.ts (6 consumers migrated to files/safe.js)

All 15 server + 3 test import sites updated to point at the canonical
location. No import goes through a shim anymore.

### Final server/utils/files/ layout

```
files/
  index.ts              ← barrel re-export
  atomic.ts             ← writeFileAtomic / writeFileAtomicSync
  safe.ts               ← statSafe / readDirSafe / readTextOrNull / resolveWithinRoot
  json.ts               ← loadJsonFile / saveJsonFile / writeJsonAtomic / readJsonOrNull
  workspace-io.ts       ← readWorkspaceText / writeWorkspaceText / readWorkspaceJson / etc.
  image-store.ts        ← saveImage / loadImageBase64
  markdown-store.ts     ← saveMarkdown / loadMarkdown
  spreadsheet-store.ts  ← saveSpreadsheet / overwriteSpreadsheet / isSpreadsheetPath
```

All tests pass, typecheck + lint + build green.
Zero stale imports confirmed via grep.
