+++ server/api/routes/scheduler.ts
@@ -1,6 +1,6 @@
 import { Router, Request, Response } from "express";
 import { WORKSPACE_PATHS } from "../../workspace/paths.js";
-import { loadJsonFile, saveJsonFile } from "../../utils/file.js";
+import { loadJsonFile, saveJsonFile } from "../../utils/files/index.js";
 import {
   dispatchScheduler,
   type SchedulerActionInput,
