+++ server/api/routes/sessions.ts
@@ -5,7 +5,7 @@ import path from "path";
 import { workspacePath } from "../../workspace/workspace.js";
 import { WORKSPACE_PATHS } from "../../workspace/paths.js";
 import { readManifest } from "../../workspace/chat-index/indexer.js";
-import { resolveWithinRoot } from "../../utils/fs.js";
+import { resolveWithinRoot } from "../../utils/files/safe.js";
 import type { ChatIndexEntry } from "../../workspace/chat-index/types.js";
 import { markRead, getSession } from "../../events/session-store/index.js";
 import { notFound } from "../../utils/httpError.js";
