+++ server/utils/files/image-store.ts
@@ -1,8 +1,8 @@
 import fs from "fs/promises";
 import path from "path";
 import crypto from "crypto";
-import { WORKSPACE_DIRS, WORKSPACE_PATHS } from "../workspace/paths.js";
-import { resolveWithinRoot } from "./fs.js";
+import { WORKSPACE_DIRS, WORKSPACE_PATHS } from "../../workspace/paths.js";
+import { resolveWithinRoot } from "./safe.js";
 
 const IMAGES_DIR = WORKSPACE_PATHS.images;
